Add tests for more interesting import assertion module map caching scenarios
Add tests for some scenarios where the module type being part of the
module map cache key (along with the specifier) affects the outcome.
For example, if an import of a given specifier uses the wrong module
type, a later import of the same specifier should succeed if it uses
the correct type assertion.

Add a wpt script for testing some of the stranger cases where repeated
fetches to the same specifier respond with different content types.

Also add basic tests for dynamic import() with the assertions
argument to the virtual/import-assertions suite.

Bug: 1132413
Change-Id: I3ff4cde5a402d3d8c502fdb2bac133323c028254