[CompositeAfterPaint] Handle main thread scrolling change during scroll animation
In pre-CAP the situation is handled when composited scrolling layer
changes in CompositedLayerMapping during compositing assignments update.
Not sure how a mere main thread scrolling status change (e.g. adding
background-attachment:fixed to a composited scrolling scroller) is
handled, but a test for the situation passes.

For CAP, now check change of main thread scrolling status in
PaintPropertyTreeBuilder.

The original test for the situation has not been working since we
removed rounded corner as a blocker of composited scrolling. Fix and
modernize it and add wpt tests.

Bug: 1176543
Change-Id: I00595635a7b1a729ddd05173e3452e40c34e52b3